requirements.txt
aiofiles==23.2.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.5
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
djangorestframework==3.14.0
kombu==5.3.3
mypy==1.7.0
mypy-extensions==1.0.0
prompt-toolkit==3.0.40
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.9
